+++ backend/Users/serializers.py
@@ -4,7 +4,6 @@ from django.contrib.auth import authenticate
 # Create serializers here :)
 from rest_framework import serializers
 from .models import User
-from Gallery.models import Photo
 from django.conf import settings
 from datetime import datetime
 from Gallery.serializers import AlbumSerializer, PhotoSerializer, CommentSerializer
@@ -62,9 +61,7 @@ class UserSerializer(serializers.ModelSerializer):
         return instance
 
 class UserPhotoSerializer(serializers.ModelSerializer):
-    #TODO: GET CONTEXT
     photos = PhotoSerializer(many = True)
-
     class Meta:
         model = User
         fields=('photos',)
