--- frontend/src/pages/User/Profile/styles.css
@@ -1,42 +1,0 @@
-.card.user-dashboard-card {
-  border: none;
-}
-.user-dashboard-card img {
-  border-radius: 50%;
-  width: 100%;
-}
-.edit-profile-dropdown .dropdown-item {
-  display: inline-block;
-}
-.edit-profile-dropdown .dropdown-item:hover {
-  background-color: var(--leit-gray-disabled);
-}
-.edit-profile label{
-  text-align: left;
-}
-
-/* User not available */
-
-.userNotAvailable {
-  min-height: 75vh;
-  text-align: center;
-}
-
-.userNotAvailable.container {
-  padding-top: 25vh;
-}
-
-.userNotAvailable h2 {
-  color: var(--leit-pink);
-  font-size: 3em;
-}
-
-.userNotAvailable p {
-  font-size: 1.5em;
-}
-
-@media (min-width: 576px) {
-  .userNotAvailable .row .col-sm-6:last-child {
-    text-align: left;
-  }
-}
