+++ frontend/src/pages/User/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "UserRouting.jsx"
+  "main": "DashboardRouting.jsx"
 }
